Add basic setup file to install client package
We use this to fetch dependencies and install the in-toto grafeas
client scripts, so that they are available on the path.

A demo can simply run
`pip install git+https://github.com/in-toto/totoify-grafeas`
and use the commands `grafeas-load`, `grafeas-run` and
`grafeas-verify` to push a layout to a grafeas server (*load),
create in-toto links and push them to a grafeas server (*run) and
fetch all metadata and verify the supply chain from that server
(*verify).