- Rename test, link -> occurrence - Fix constructor - Add grafeas style artifact objects - Handle return values - Misc fixes
If we convert all byproducts and environment fields to str, we have no path back.
It's likely better to let an in-toto to Grafeas conversion fail because of
disallowed types, as then the user will know what's supported. A blanket conversion
to str() could mean a user proceeds with the conversion with no way to revert and
verify metadata.

Signed-off-by: Aditya Sirish <aditya@saky.in>
Signed-off-by: Kristel Fung <kristelfung@gmail.com>